[refactor] admin 회사 정보 api 호출 함수로 변경
