+++ src/components/admin/index.ts
@@ -1,2 +1,3 @@
 export { AdminProductDialog } from './AdminProductDialog';
+export { CompanyInfo } from './CompanyInfo';
 export { SideBar } from './SideBar';
